feat(mobile): android terminals inbox and new-terminal sheet
