+++ apps/mobile/src/components/terminal/new-terminal-sheet.types.ts
@@ -1,0 +1,7 @@
+export interface NewTerminalSheetProps {
+  isPresented: boolean;
+  onIsPresentedChange: (isPresented: boolean) => void;
+  creating: boolean;
+  error: string | null;
+  onCreate: (cwd: string) => Promise<boolean>;
+}
